Create Api class
