+++ src/index.js
@@ -1,0 +1,11 @@
+import Api from 'Api'
+
+
+const api = new Api({
+  baseUrl: 'https://nomoreparties.co/v1/plus-cohort-21',
+  headers: {
+    authorization: 'f07db228-d9f6-44b3-81fa-e5425e2f6a35',
+    'Content-Type': 'application/json; charset=UTF-8'
+  }
+});
+
